+++ core/src/main/java/com/refinedmods/refinedstorage2/core/storage/Storage.java
@@ -6,11 +6,11 @@ import java.util.Optional;
 import com.refinedmods.refinedstorage2.core.util.Action;
 
 public interface Storage<T> {
-    Optional<T> extract(T template, int amount, Action action);
+    Optional<T> extract(T template, long amount, Action action);
 
-    Optional<T> insert(T template, int amount, Action action);
+    Optional<T> insert(T template, long amount, Action action);
 
     Collection<T> getStacks();
 
-    int getStored();
+    long getStored();
 }
