+++ core/src/main/java/com/refinedmods/refinedstorage2/core/storage/disk/StorageDisk.java
@@ -3,5 +3,5 @@ package com.refinedmods.refinedstorage2.core.storage.disk;
 import com.refinedmods.refinedstorage2.core.storage.Storage;
 
 public interface StorageDisk<T> extends Storage<T> {
-    int getCapacity();
+    long getCapacity();
 }
